% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc}
\alias{bc}
\alias{bc.spots}
\title{Human Breast Cancer (BC) Dataset}
\format{
\code{bc} is a matrix with 250 spots and 14,789 genes that denotes
the expression count table. \code{bc.spots} is a matrix of the
geospatial profile.
}
\source{
Publicly available at \url{http://www.spatialresearch.org}.
}
\description{
The human breast cancer (BC) study is a publicly available ST dataset
accessible on the Spatial Research Lab \url{http://www.spatialresearch.org}.
There are four layers in the BC study. Following the previous
studies (Li et al., 2020; Sun et al., 2020; Svensson et al., 2018), the
BC layer 2, which contains 14,789 genes measured on 250 spots, was saved.
}
\examples{
## human breast cancer (BC) dataset
data(bc)

}
\references{
Li, Q., M. Zhang, Y. Xie, and G. Xiao (2020). Bayesian modeling of
spatial molecular profiling data via Gaussian process.
\emph{arXiv preprint arXiv:2012.03326}.

Sun, S., J. Zhu, and X. Zhou (2020). Statistical analysis of spatial
expression patterns for spatially resolved transcriptomic studies.
\emph{Nature Methods 17} (2), 193–200.

Svensson, V., S. A. Teichmann, and O. Stegle (2018).
SpatialDE: Identification of spatially variable genes.
\emph{Nature Methods 15} (5), 343–346.
}
\keyword{dataset}
