% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinSpect.R
\name{binSpect}
\alias{binSpect}
\title{Binary Spatial Extraction of Genes}
\usage{
binSpect(bin.expr, neighbor.info, do.fisher.test = FALSE, gene.name = NULL)
}
\arguments{
\item{bin.expr}{A numeric vector \eqn{p} of length \eqn{n} that denotes the
dichotomised gene expression levels. Each entry is one if the gene is
highly expressed at spot \eqn{i} and zero otherwise.}

\item{neighbor.info}{An \eqn{n}-by-\eqn{K} numeric matrix \eqn{A} that
denotes the long format of the adjacency matrix. Each entry denotes the
neighbor for spot \eqn{i}.}

\item{do.fisher.test}{A logical value that indicates if the Fisher test
should be used. The default value is \code{FALSE} to use a normal
approximation test.}

\item{gene.name}{A character string that specifies the name of the gene
passed. To be used when storing the results. The default value is \code{NULL}
to keep the gene expression levels unnamed.}
}
\value{
\code{binSpect} returns an object of class "\code{binSpect}".
The function \code{\link[base:print]{base::print()}} i.e., \code{\link[=print.binSpect]{print.binSpect()}}, can be used to
print a summary of the results.

An object of class "\code{binSpect}" is a list containing the following components:

\item{call}{the function call in which all of the specified arguments are specified by their full names.}
\item{model}{the name of statistical model or technique.}
\item{gene.name}{the name of gene evaluated.}
\item{summary}{a summary table that contains the contingency table of low-expressed and highly-expressed interactions.}
\item{measures}{the estimated odds ratio and corresponding \eqn{p}-value.}
\item{time}{the execution time of the function.}
}
\description{
BinSpect is a fast computational method that identifies spatially variable
genes using the binarized expression levels.
}
\details{
This method uses a graph-based approach to compute the probability of
encountering two physically neighboring cells being both
highly-expressed (=1). It is based on the provided neighborhood information,
which can be interpreted as a graph structure. Based on the dichotomised
gene expression levels, an edge in the "graph" connecting two cells is
classified as 1-0, 1-1, 0-1, or 0-0. A contingency table that counts
all the classified edges for the entire neighborhood network is created.
A \eqn{p}-value is then reported based on the hypergeometric
distribution test.

See \url{http://spatialgiotto.rc.fas.harvard.edu/giotto_spatial_genes.html}
for more information.
}
\examples{
## Need to implement this example.

}
\references{
Dries, R., Zhu, Q., Dong, R. et al. Giotto: a toolbox for
integrative analysis and visualization of spatial expression data.
\emph{Genome Biol} \strong{22}, 78 (2021).
\url{https://doi.org/10.1186/s13059-021-02286-2}
}
\seealso{
\code{\link[=binarize.st]{binarize.st()}} for dichotomising the expression levels;
\code{\link[=get.neighbors]{get.neighbors()}} for getting the neighborhood information.
}
\keyword{method}
