\name{Boost-package}
\alias{Boost-package}
\alias{Boost}
\docType{package}
\title{
\packageTitle{Boost}
}
\description{
\packageDescription{Boost}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Boost}
\packageIndices{Boost}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Boost}

Maintainer: \packageMaintainer{Boost}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
