% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOST.GP.R
\name{BOOST.GP}
\alias{BOOST.GP}
\title{Fit BOOST-GP Model for Gene Expression Levels}
\usage{
BOOST.GP(
  abs.expr,
  spots,
  size.factor = NULL,
  gene.name = NULL,
  n.iter = 10000,
  burn.prop = 0.5,
  update.prop = 0.2,
  init.b.sigma = NULL,
  init.h = 1
)
}
\arguments{
\item{abs.expr}{A numeric vector \eqn{p} of length \eqn{n} that denotes the
absolute gene expression levels. Each entry is an integer that denotes
the gene count at spot \eqn{i}.}

\item{spots}{An \eqn{n}-by-\eqn{2} numeric matrix \eqn{T} to represent the
geospatial profile, where each row indicates the spot location in
the grid.}

\item{size.factor}{A numeric vector \eqn{s} of length \eqn{n} to compute
the relative gene expression levels. Each entry denotes the size factor
of sample \eqn{i} that captures all nuisance effects. The default
is \code{NULL} to evaluate the absolute expression levels.}

\item{gene.name}{A character string that specifies the name of the gene
passed. To be used when storing the results. The default value is \code{NULL}
to keep the gene expression levels unnamed.}

\item{n.iter}{An integer value to specify the number of iterations for the
DMH algorithm. The default is 10,000 iterations.}

\item{burn.prop}{A numeric value to specify the proportion of iterations to
use as warm-up. The default is 0.50 to use half of the iterations
for warm-up.}

\item{update.prop}{A numeric value to specify the proportion of samples to
update in each iteration. The default is 0.2 to update one-fifth of the
total samples.}

\item{init.b.sigma}{A numeric value to specify the initial value of the
scale parameter in the inverse-gamma prior on the variance of the
multivariate normal distribution prior for the log-expression levels.
The default is \code{NULL} to set the value as twice the sample variance of
the log-expression levels.}

\item{init.h}{A numeric value to specify the scaling of the variance for
the normal prior set on each coefficient. The default is one to not scale
the variance.}
}
\value{
\code{BOOST.GP} returns an object of class "\code{BOOST.GP}".
The function \code{\link[base:print]{base::print()}} i.e., \code{\link[=print.BOOST.GP]{print.BOOST.GP()}}, can be used to
print a summary of the results.

An object of class "\code{BOOST.GP}" is a list containing the following components:

\item{call}{the function call in which all of the specified arguments are specified by their full names.}
\item{model}{the name of statistical model or technique.}
\item{gene.name}{the name of gene evaluated.}
\item{summary}{a summary table that contains a summary of the estimated parameters.}
\item{measures}{the estimated Bayes factor and corresponding \eqn{p}-value}
\item{time}{the execution time of the function.}
}
\description{
Fit the BOOST-GP model to detect whether the gene is spatially
variable (SV). The fit is done within a Metropolis (SSVS) search variable
selection algorithm. Only one gene must be present, but no normalization
is necessary.
}
\details{
The primary interest lies in the identification of SV genes via a
selection indicator. See Li et al. (2020) for more information on the
model fitting and posterior inference procedures.
}
\examples{
\dontrun{
library(boost)

## load sample dataset
data(mob)

## extract a sample gene and get size factor
g <- mob[, "Apoe"]
s <- get.size.factor(mob, estimation.method = "TSS")

## fit the model
res <- BOOST.GP(g, mob.spots, size.factor = s, gene.name = "Apoe", n.iter = 200)
print(res)
}
}
\references{
Li, Q., Zhang M., Xie Y., & Xiao, G. (2020). Bayesian Modeling
of Spatial Molecular Profiling Data via Gaussian Process.
\emph{arXiv preprint arXiv:2012.03326}.
}
\seealso{
\code{\link[=get.size.factor]{get.size.factor()}} for estimating the size factor;
\code{\link[=print.BOOST.GP]{print.BOOST.GP()}} for printing a summary of results to console.
}
\keyword{BOOST-GP}
\keyword{method}
