% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BooSTmore.R
\name{BooST.more}
\alias{BooST.more}
\title{Adds more trees to an estimated BooST object}
\usage{
BooST.more(x, y, object, M, display = FALSE)
}
\arguments{
\item{x}{Design matrix with explanatory variables.}

\item{y}{Response variable.}

\item{M}{Number of trees.}

\item{display}{If TRUE, displays iteration counter.}
}
\value{
An object with S3 class "Boost".
\item{Model}{A list with all trees.}
\item{fitted.values}{Final model fitted values.}
\item{brmse}{Boost rmse in each iteratiob.}
\item{Model}{A list with all trees.}
\item{ybar}{Average value of y used in the first iteration.}
\item{v}{Chosen learning rate.}
\item{rho}{Vector of gradient estimates for each iteration.}
\item{nvar}{Numver of variables in x}
\item{varnames}{colnames of x to be used in other functions.}
\item{call}{The matched call.}
}
\description{
Adds more trees to an estimated BooST object
}
\examples{
## == to be made == ##

}
\references{
blablabla
}
\keyword{BooST,}
\keyword{Boosting,}
\keyword{Effects}
\keyword{Partial}
\keyword{Smooth}
\keyword{Tree,}
