% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost-package.R
\docType{package}
\name{boost}
\alias{boost}
\title{boost: Bayesian Modeling of Spatial Transcriptomics Data}
\description{
This package provides functions to detect genes with spatial expression
pattern, also known as spatially variable (SV) genes, in spatial
transcriptomics (ST) data. Two major and novel Bayesian models are
implemented via a Gaussian process or Ising model. In addition, it also
provides other standard statistical tools such as SPARK, binSpect, etc.
Utilities are available to normalize count data, dichotomise expression
levels, get spatial neighbors, and view the results of the procedures.
}
\details{
See Jiang et al. (2021) and Li et al. (2020) for details and illustrations
of how the Bayesian models are fitted and their applications.
}
\references{
Jiang, X., Li, Q., & Xiao, G. (2021). Bayesian Modeling of
Spatial Transcriptomics Data via a Modified Ising Model.
\emph{arXiv preprint arXiv:2104.13957}.

Li, Q., Zhang M., Xie Y., & Xiao, G. (2020). Bayesian Modeling of
Spatial Molecular Profiling Data via Gaussian Process.
\emph{arXiv preprint arXiv:2012.03326}.
}
