% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute.metrics}
\alias{compute.metrics}
\title{Performance of Gene Identification}
\usage{
compute.metrics(
  predictor,
  truth,
  predictor.type = c("BF", "p-value"),
  threshold = NULL
)
}
\arguments{
\item{predictor}{A numeric vector of length \eqn{n} that denotes the
\eqn{p}-values or Bayes factors (BFs).}

\item{truth}{A logical vector of length \eqn{n} that represents the ground
truth corresponding to the predictor.}

\item{predictor.type}{A character string that specifies whether
\eqn{p}-values of Bayes factors (BFs) were provided.}

\item{threshold}{A numeric value that specifies the cutoff for
defining SV genes.}
}
\value{
A list object that contains six performance metrics
(Sensitivity, Specificity, F1_score, FDR, AUC, and MCC).
}
\description{
Calculate the performance of spatially variable (SV) gene identification
on simulated data.
}
\keyword{metrics}
