% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative_expression.R
\name{derivative_expression}
\alias{derivative_expression}
\title{Creates analytical expression for partial effects}
\usage{
derivative_expression(object)
}
\arguments{
\item{object}{A BooST or a SmoothTree object.}
}
\description{
Creates analytical expression for partial effects estimation of a BoosT or a SmoothTree object.
}
\examples{
## == to be made == ##

}
\keyword{BooST,}
\keyword{Boosting,}
\keyword{Effects,}
\keyword{Partial}
\keyword{Predict}
\keyword{Smooth}
\keyword{Tree,}
