% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_derivative.R
\name{estimate_derivatives}
\alias{estimate_derivatives}
\title{Creates analytical expression for partial effects}
\usage{
estimate_derivatives(object, x, de_exp)
}
\arguments{
\item{object}{A BooST or a SmoothTree object.}

\item{x}{Matrix of obsevations where the partial derivatives are desired.}

\item{de_exp}{Derivative expressions from the derivative_expression function.}
}
\description{
Creates analytical expression for partial effects estimation of a BoosT or a SmoothTree object.
}
\examples{
## == to be made == ##

}
\keyword{BooST,}
\keyword{Boosting,}
\keyword{Effects}
\keyword{Partial}
\keyword{Smooth}
\keyword{Tree,}
