% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.factor.R
\name{get.size.factor}
\alias{get.size.factor}
\title{Get Size Factor Estimate}
\usage{
get.size.factor(count, estimation.method = c("TSS", "Q75", "RLE", "TMM"))
}
\arguments{
\item{count}{An \eqn{n}-by-\eqn{p} numeric matrix \eqn{Y} that denotes the
absolute gene expression count table. Each entry is the absolute read
count for gene \eqn{j} collected at spot \eqn{i}.}

\item{estimation.method}{An optional character string to specify the
size factor computation technique. The default is "TSS" for
total sum scaling.}
}
\value{
A numeric vector, where each entry denotes the size factor for
sample \eqn{i}.
}
\description{
Obtain the size factor estimates to capture all nuisance effects and
compute the relative gene expression levels.
}
\details{
Normalization is critical to sequence count data analysis. As a result, to
counteract various artifacts and bias due to biological and technical
reasons, the read counts should be converted to their relative gene
expression levels i.e., divide the counts by the result of this function.
If the main interest is in the absolute gene expression level, then
the size factor is set to the same value; otherwise, the size factor is
computed directly from the gene expression count data.

This method offers the following options for computing size factors:
total sum scaling (TSS), upper-quartiles (Q75), relative log
expression (RLE), and weighted trimmed mean by M-values (TMM).

See Jiang et al. (2021) for more information and references on
these options.
}
\references{
Jiang, X., Li, Q., & Xiao, G. (2021). Bayesian Modeling of
Spatial Transcriptomics Data via a Modified Ising Model.
\emph{arXiv preprint arXiv:2104.13957}.
}
\keyword{preprocessing}
