% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mob}
\alias{mob}
\alias{mob.spots}
\title{Mouse Olfactory Bulb (MOB) Dataset}
\format{
\code{mob} is a matrix with 262 spots and 16,218 genes that denotes
the expression count table. \code{mob.spots} is a matrix of the
geospatial profile.
}
\source{
Publicly available at \url{http://www.spatialresearch.org}.
}
\description{
The mouse olfactory bulb (MOB) study is a publicly available ST dataset
accessible on the Spatial Research Lab \url{http://www.spatialresearch.org}.
There are twelve replicates in the MOB study. Following the previous
studies (Li et al., 2020; Sun et al., 2020; Svensson et al., 2018), the
MOB replicate eleven, which contains 16,218 genes measured on 262 spots,
was saved.
}
\examples{
## mouse olfactory bulb (MOB) dataset
data(mob)

}
\references{
Li, Q., M. Zhang, Y. Xie, and G. Xiao (2020). Bayesian modeling of
spatial molecular profiling data via Gaussian process.
\emph{arXiv preprint arXiv:2012.03326}.

Sun, S., J. Zhu, and X. Zhou (2020). Statistical analysis of spatial
expression patterns for spatially resolved transcriptomic studies.
\emph{Nature Methods 17} (2), 193–200.

Svensson, V., S. A. Teichmann, and O. Stegle (2018).
SpatialDE: Identification of spatially variable genes.
\emph{Nature Methods 15} (5), 343–346.
}
\keyword{dataset}
