% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize.st}
\alias{normalize.st}
\title{Normalize Sequence Count Data}
\usage{
normalize.st(
  count,
  scaling.method = c("TSS", "Q75", "RLE", "TMM", "A-VST", "N-VST", "log-VST")
)
}
\arguments{
\item{count}{An \eqn{n}-by-\eqn{p} integer matrix \eqn{Y} that denotes the
absolute gene expression count table. Each entry is the read count for
gene \eqn{j} collected at spot \eqn{i}.}

\item{scaling.method}{An optional character string to specify the
normalization technique. The default is "TSS" for the for total sum
scaling.}
}
\value{
A numeric matrix that denotes the relative expression levels.
}
\description{
Obtain the relative expression levels for the gene expression count table.
}
\details{
Normalization is critical to the development of analysis techniques for any
sequence count data that suffer from various sequence artifacts and biases.
This method provides eight choices of normalization methods in two
categories that attempt to counteract those biases due to biological and
technical reasons. The first type is based on size factor estimation
(i.e., TSS, Q75, RLE, TMM) where a quantity is obtained, for each
sample, that captures all nuisance effects. The other type of normalization
method  is based on variance-stabilizing transformation (VST)
(i.e., A-VST, N-VST, log-VST), which aims to transform a random variable
with a negative binomial distribution into one with an approximately
normal distribution.

See Jiang et al. (2021) for more information on the normalization methods.
}
\references{
Jiang, X., Li, Q., & Xiao, G. (2021). Bayesian Modeling of
Spatial Transcriptomics Data via a Modified Ising Model.
\emph{arXiv preprint arXiv:2104.13957}.
}
\seealso{
\code{\link[=st.plot]{st.plot()}} for plotting the relative expression levels.
}
\keyword{preprocessing}
