% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_smoothtree.R
\name{predict.SmoothTree}
\alias{predict.SmoothTree}
\title{Predict Method for Smooth Tree Fits}
\usage{
\method{predict}{SmoothTree}(object, newx = NULL, ...)
}
\arguments{
\item{object}{A BooST or a SmoothTree object.}

\item{newx}{Matrix of new values of x at wich predictions are to be made.}

\item{...}{Additional arguments for other methods.}
}
\description{
Obtains predictions for a Smooth Tree model and a given design matrix.
}
\examples{
## == to be made == ##

}
\keyword{BooST,}
\keyword{Boosting,}
\keyword{Effects,}
\keyword{Partial}
\keyword{Predict}
\keyword{Smooth}
\keyword{Tree,}
