% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinSpect.R
\name{print.binSpect}
\alias{print.binSpect}
\title{Print BinSpect Results}
\usage{
\method{print}{binSpect}(x, ...)
}
\arguments{
\item{x}{An object of class \code{binSpect}.}

\item{...}{Additional arguments passed to \code{\link[base:print]{base::print()}}.}
}
\description{
Print BinSpect Results
}
