% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOST.GP.R
\name{print.BOOST.GP}
\alias{print.BOOST.GP}
\title{Print BOOST-GP Model Fitting Results}
\usage{
\method{print}{BOOST.GP}(x, ...)
}
\arguments{
\item{x}{An object of class \code{BOOST.GP}.}

\item{...}{Additional arguments passed to \code{\link[base:print]{base::print()}}.}
}
\value{
No return value, called for side effects i.e. printing to console.
}
\description{
S3 class method printing MCMC results from BOOST-GP fitted model. The
console displays the function call, model type, parameter estimates, and
Bayes Factor (BF) measurement for testing wether the gene is spatially
variable.
}
\examples{
## See example given in function for BOOST-GP

}
\seealso{
\code{\link[=BOOST.GP]{BOOST.GP()}} for BOOST-GP model fitting.
}
\keyword{BOOST-GP}
\keyword{diagnostics}
