% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOOST.Ising.R
\name{print.BOOST.Ising}
\alias{print.BOOST.Ising}
\title{Print BOOST-Ising Model Fitting Results}
\usage{
\method{print}{BOOST.Ising}(x, ...)
}
\arguments{
\item{x}{An object of class \code{BOOST.Ising}.}

\item{...}{Additional arguments passed to \code{\link[base:print]{base::print()}}.}
}
\value{
No return value, called for side effects i.e. printing to console.
}
\description{
S3 class method printing MCMC results from BOOST-Ising fitted model. The
console displays the function call, model type, parameter estimates, and
Bayes Factor (BF) measurement for testing an attraction pattern in the gene.
}
\examples{
## Need to implement the example for the procedure.

}
\seealso{
\code{\link[=BOOST.Ising]{BOOST.Ising()}} for BOOST-Ising model fitting.
}
\keyword{BOOST-Ising}
\keyword{diagnostics}
