% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARK.R
\name{print.SPARK}
\alias{print.SPARK}
\title{Print SPARK Results}
\usage{
\method{print}{SPARK}(x, ...)
}
\arguments{
\item{x}{An object of class \code{SPARK}.}

\item{...}{Additional arguments passed to \code{\link[base:print]{base::print()}}.}
}
\description{
Print SPARK Results
}
