% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{print.SpatialDE}
\alias{print.SpatialDE}
\title{Print SpatialDE Results}
\usage{
\method{print}{SpatialDE}(x, ...)
}
\arguments{
\item{x}{An object of class \code{SpatialDE}.}

\item{...}{Additional arguments passed to \code{\link[base:print]{base::print()}}.}
}
\description{
Print SpatialDE Results
}
