% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARK.R
\name{SPARK}
\alias{SPARK}
\title{Spatially Resolved Transcriptomic Analysis}
\usage{
SPARK(abs.expr, spots, size.factor, gene.name = NULL)
}
\arguments{
\item{abs.expr}{A numeric vector \eqn{p} of length \eqn{n} that denotes the
absolute gene expression levels. Each entry is an integer that denotes
the gene count at spot \eqn{i}.}

\item{spots}{An \eqn{n}-by-\eqn{2} numeric matrix \eqn{T} to represent the
geospatial profile, where each row indicates the spot location in
the grid.}

\item{size.factor}{A numeric vector \eqn{s} of length \eqn{n} to compute
the relative gene expression levels. Each entry denotes the size factor
of sample \eqn{i} that captures all nuisance effects.}

\item{gene.name}{A character string that specifies the name of the gene
passed. To be used when storing the results. The default value is \code{NULL}
to keep the gene expression levels unnamed.}
}
\value{
\code{SPARK} returns an object of class "\code{SPARK}".
The function \code{\link[base:print]{base::print()}} i.e., \code{\link[=print.SPARK]{print.SPARK()}}, can be used to
print a summary of the results.

An object of class "\code{SPARK}" is a list containing the following components:
\item{call}{the function call in which all of the specified arguments are specified by their full names.}
\item{model}{the name of statistical model or technique.}
\item{gene.name}{the name of gene evaluated.}
\item{summary}{a summary table that contains the p-values for the different tests.}
\item{measures}{the adjusted and combined \eqn{p}-values.}
\item{time}{the execution time of the function.}
}
\description{
SPARK is an efficient method to identify spatially-variable genes.
}
\details{
This method directly models count data generated from various spatial
resolved transcriptomic techniques through generalized spatial
linear models. It relies on penalized quasi-likelihood algorithm for
scalable computation and recently developed statistical formulas for
hypothesis testing, providing effective control of type I errors and
yielding high statistical power.

See \url{https://xzhoulab.github.io/SPARK/} for more information.
}
\examples{
## Need to implement this example.

}
\references{
Sun, S., Zhu, J. & Zhou, X. Statistical analysis of
spatial expression patterns for spatially resolved transcriptomic studies.
\emph{Nat Methods} \strong{17}, 193–200 (2020).
\url{https://doi.org/10.1038/s41592-019-0701-7}.
}
\seealso{
\code{\link[=get.size.factor]{get.size.factor()}} for obtaining the size factors.
}
\keyword{method}
