% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDE.R
\name{SpatialDE}
\alias{SpatialDE}
\title{Statistical Test to Identify Genes with Spatial Patterns}
\usage{
SpatialDE(norm.expr, spots, gene.name = NULL)
}
\arguments{
\item{norm.expr}{A numeric vector \eqn{p} of length \eqn{n} that denotes the
relative gene expression levels. Each entry is a numeric value that
denotes the normalized gene count at spot \eqn{i}.}

\item{spots}{An \eqn{n}-by-\eqn{2} numeric matrix \eqn{T} to represent the
geospatial profile, where each row indicates the spot location in
the grid.}

\item{gene.name}{A character string that specifies the name of the gene
passed. To be used when storing the results. The default value is \code{NULL}
to keep the gene expression levels unnamed.}
}
\value{
\code{SpatialDE} returns an object of class "\code{SpatialDE}".
The function \code{\link[base:print]{base::print()}} i.e., \code{\link[=print.SpatialDE]{print.SpatialDE()}}, can be used to
print a summary of the results.

An object of class "\code{SpatialDE}" is a list containing the following components:

\item{call}{the function call in which all of the specified arguments are specified by their full names.}
\item{model}{the name of statistical model or technique.}
\item{gene.name}{the name of gene evaluated.}
\item{summary}{a summary table that contains a summary of the model.}
\item{measures}{the estimated \eqn{p}-value.}
\item{time}{the execution time of the function.}
}
\description{
SpatialDE is a method to identify and characterize spatially variable (SV)
genes via a statistical test.
}
\details{
This method builds on Gaussian process regression, a class of models.
For a given gene, the expression variability is decomposed into spatial and
non-spatial components, using two random effect terms: a spatial variance
term that parametrizes gene expression covariance by pairwise distances of
samples, and a noise term that models non-spatial variability. The ratio
of the variance explained by these two components quantifies the
Fraction of Spatial Variance (FSV). Significant SV genes can be identified
by comparing this full model to a model without the
spatial variance component.

See \url{https://bioconductor.org/packages/release/bioc/html/spatialDE.html}
for more information.
}
\examples{
library(boost)

## load sample dataset
data(mob)

## scale gene expression levels and extract one gene
mob.t <- normalize.st(mob, scaling.method = "TSS")
g     <- mob.t[, "Apoe"]

## run the statistical test
res <- SpatialDE(g, mob.spots, gene.name = "Apoe")
print(res)
}
\references{
Corso D, Malfait M, Moses L (2021). \emph{spatialDE: R wrapper for SpatialDE}.
doi: \href{https://doi.org/10.18129/B9.bioc.spatialDE}{10.18129/B9.bioc.spatialDE},
R package version 1.0.0, \url{http://www.bioconductor.org/packages/spatialDE}.

Svensson V, Teichmann SA, Stegle O (2018).
"SpatialDE: identification of spatially variable genes."
\emph{Nature Methods}, \strong{15}(5), 343–346. ISSN 1548-7105,
doi: \href{https://doi.org/10.1038/nmeth.4636}{10.1038/nmeth.4636},
\url{https://www.nature.com/articles/nmeth.4636}.
}
\seealso{
\code{\link[=normalize.st]{normalize.st()}} for normalizing the expression levels;
\code{\link[=print.SpatialDE]{print.SpatialDE()}} for printing a summary of results to console.
}
\keyword{method}
