% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{st.plot}
\alias{st.plot}
\title{Spatial Transcriptomics Plot}
\usage{
st.plot(expr, spots, gene.name = NULL, binary = FALSE, log.expr = TRUE)
}
\arguments{
\item{expr}{A numeric vector \eqn{p} of length \eqn{n} that denotes the
absolute, dichotomised, or relative gene expression levels. Each entry is
an integer value that denotes the gene count at spot \eqn{i}.}

\item{spots}{An \eqn{n}-by-\eqn{2} numeric matrix \eqn{T} to represent the
geospatial profile, where each row indicates the spot location in
the grid.}

\item{gene.name}{An optional character string that specifies the gene name.
The default is \code{NULL} which appends nothing to the plot title.}

\item{binary}{A logical value that indicates if the gene expression levels
are dichotomised. The default is \code{FALSE} to expect absolute or relative
expression levels.}

\item{log.expr}{A logical value that indicates if the gene expression levels
should be plotted at log-level. The default is \code{TRUE} to log-transform
the gene expression levels if they are not dichotomised.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the absolute, dichotomised, or relative gene expression levels in a
spatial transcriptomics (ST) dataset.
}
\examples{
## Need to implement this example.

}
\keyword{plot}
