% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BooST.R
\name{BooST}
\alias{BooST}
\title{Estimate BooST}
\usage{
BooST(x, y, v = 0.2, p = 2/3, d_max = 4, gamma = NULL, M = 300,
  display = FALSE, stochastic = FALSE, s_prop = 0.5)
}
\arguments{
\item{x}{Design matrix with explanatory variables.}

\item{y}{Response variable.}

\item{v}{Learning rate (default 0.2).}

\item{p}{Proportion of variables tested in each node split (default 2/3).}

\item{d_max}{Number of nodes in each tree (default 4). If NULL each tree will have a random number of nodes between 4 and 8.}

\item{gamma}{aa}

\item{M}{Number of trees.}

\item{display}{If TRUE, displays iteration counter.}

\item{stochastic}{If TRUE the model will be estimated using Stochasting Gradient Boosting.}

\item{s_prop}{Used only if stochastic=TRUE. Determines the proportion of data used in each tree.}
}
\value{
An object with S3 class "Boost".
\item{Model}{A list with all trees.}
\item{fitted.values}{Final model fitted values.}
\item{brmse}{Boost rmse in each iteratiob.}
\item{Model}{A list with all trees.}
\item{ybar}{Average value of y used in the first iteration.}
\item{v}{Chosen learning rate.}
\item{rho}{Vector of gradient estimates for each iteration.}
\item{nvar}{Numver of variables in x}
\item{varnames}{colnames of x to be used in other functions.}
\item{call}{The matched call.}
}
\description{
Estimates Boosting of Smooth Trees (BooST)
}
\examples{
## == to be made == ##

}
\references{
blablabla
}
\keyword{BooST,}
\keyword{Boosting,}
\keyword{Effects}
\keyword{Partial}
\keyword{Smooth}
\keyword{Tree,}
