% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BooST.R
\name{BooST}
\alias{BooST}
\title{Estimate BooST}
\usage{
BooST(
  x,
  y,
  v = 0.2,
  p = 2/3,
  d_max = 4,
  gamma = seq(0.5, 5, 0.01),
  M = 300,
  display = FALSE,
  stochastic = FALSE,
  s_prop = 0.5,
  node_obs = nrow(x)/200,
  random = FALSE
)
}
\arguments{
\item{x}{Design matrix with explanatory variables.}

\item{y}{Response variable.}

\item{v}{Learning rate (default 0.2).}

\item{p}{Proportion of variables tested in each node split (default 2/3).}

\item{d_max}{Number of splits in each tree (default 4).}

\item{gamma}{Transiction function intensity. Bigger numbers makes the transition less smoth. The default is a sequence of values (0.5:5) to be randomized in each new node. Multiple values may be supplied in a vector to increase the model randomness.}

\item{M}{Number of trees.}

\item{display}{If TRUE, displays iteration counter.}

\item{stochastic}{If TRUE the model will be estimated using Stochasting Gradient Boosting.}

\item{s_prop}{Used only if stochastic=TRUE. Determines the proportion of data used in each tree.}

\item{node_obs}{Equivalent to the minimum number of observations in a termina node for a discrete tree.}

\item{random}{If TRUE trees are grown randomly (default = FALSE)}
}
\value{
An object with S3 class "Boost".
\item{Model}{A list with all trees.}
\item{fitted.values}{Final model fitted values.}
\item{brmse}{Boost rmse in each iteratiob.}
\item{Model}{A list with all trees.}
\item{ybar}{Average value of y used in the first iteration.}
\item{v}{Chosen learning rate.}
\item{rho}{Vector of gradient estimates for each iteration.}
\item{nvar}{Numver of variables in x}
\item{varnames}{colnames of x to be used in other functions.}
\item{params}{Model parameters.}
\item{call}{The matched call.}
}
\description{
Estimates Boosting of Smooth Trees (BooST)
}
\examples{
## == to be made == ##

}
\references{
blablabla
}
\keyword{BooST,}
\keyword{Boosting,}
\keyword{Effects}
\keyword{Partial}
\keyword{Smooth}
\keyword{Tree,}
