% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{rank.aggregation}
\alias{rank.aggregation}
\title{Aggregate Rankings}
\usage{
rank.aggregation(
  data,
  K,
  method = c("GEO", "MC2"),
  ties.method = c("average", "first", "last", "random", "max", "min")
)
}
\arguments{
\item{data}{An \eqn{n}-by-\eqn{(p+1)} numeric matrix that represents a
table of \eqn{p}-values with \eqn{n} observations and
\eqn{p+1} variables. Note that the first column should be the gene name
and the columns following after should be the methods
(i.e., base rankers).}

\item{K}{A numeric value that indicates how many top-\eqn{K} genes to sort
out in each base ranker.}

\item{method}{A character string that specifies the rank aggregation method.}

\item{ties.method}{A character string that specifies how ties should
be treated (see "Details" in \code{\link[base:rank]{base::rank()}}).}
}
\value{
A numeric matrix that represents a table with genes and their rank.
}
\description{
Generate an aggregated ranking from multiple base rankers.
}
\details{
See Schimek, M. G. et al. (2015) for more information on the procedure.
}
\references{
Schimek M., Budinska E., Kugler K., Svendova V., Ding J., Lin S.
(2015). “TopKLists: a comprehensive R package for statistical inference,
stochastic aggregation, and visualization of multiple omics ranked lists.”
\emph{Stat Appl Genet Mol Biol}, 311-6.
\url{http://www.degruyter.com/doi/10.1515/sagmb-2014-0093}.
}
